fix:fluxo erro

+++ src/Tiradentes.CobrancaAtiva.Application/ViewModels/Cobranca/ErroParcelaViewModel.cs
@@ -5,7 +5,7 @@ namespace Tiradentes.CobrancaAtiva.Application.ViewModels.Cobranca
     {
         public int Etapa { get; set; }
         public decimal IdErro { get; set; }
-        public decimal CodErro { get; set; }
+        public int CodErro { get; set; }
         public decimal ValorParcela { get; set; }
     }
 }
