+++ src/Tiradentes.CobrancaAtiva.Domain/Models/ItensBaixaTipo1Model.cs
@@ -8,7 +8,7 @@ namespace Tiradentes.CobrancaAtiva.Domain.Models
         public DateTime DataBaixa { get; set; }
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public decimal Sequencia { get; set; }
-        public decimal? CodigoErro { get; set; }
+        public int? CodigoErro { get; set; }
         public decimal NumeroLinha { get; set; }
         public decimal NumeroAcordo { get; set; }
         public decimal Matricula { get; set; }
