+++ src/Tiradentes.CobrancaAtiva.Infrastructure/Repositories/ItensBaixasTipo1Repository.cs
@@ -25,7 +25,7 @@ namespace Tiradentes.CobrancaAtiva.Infrastructure.Repositories
             await Alterar(itemBaixa);
         }
 
-        public async Task InserirBaixa(DateTime dataBaixa, decimal matricula, decimal numeroAcordo, decimal multa, decimal juros, DateTime dataVencimento, decimal valorParcela, decimal? codErro, string cnpjEmpresaCobranca, int parcela, string sistema, string situacaoAluno, string tipoInadimplencia)
+        public async Task InserirBaixa(DateTime dataBaixa, decimal matricula, decimal numeroAcordo, decimal multa, decimal juros, DateTime dataVencimento, decimal valorParcela, int? codErro, string cnpjEmpresaCobranca, int parcela, string sistema, string situacaoAluno, string tipoInadimplencia)
         {
             HabilitarAlteracaoBaixaCobranca(true);
