+++ src/Tiradentes.CobrancaAtiva.Services/Interfaces/IItensBaixasTipo2Service.cs
@@ -5,6 +5,6 @@ namespace Tiradentes.CobrancaAtiva.Services.Interfaces
 {
     public interface IItensBaixasTipo2Service : IDisposable
     {
-        Task InserirBaixa(DateTime dataBaixa, decimal matricula, decimal numeroAcordo, int parcela, decimal periodo, DateTime dataVencimento, decimal valor, decimal? codErro);
+        Task InserirBaixa(DateTime dataBaixa, decimal matricula, decimal numeroAcordo, int parcela, decimal periodo, DateTime dataVencimento, decimal valor, int? codErro);
     }
 }
