+++ src/Tiradentes.CobrancaAtiva.Services/Services/ItensBaixasTipo1Service.cs
@@ -19,7 +19,7 @@ namespace Tiradentes.CobrancaAtiva.Services.Services
                                                  matricula);
         }
 
-        public async Task InserirBaixa(DateTime dataBaixa, decimal matricula, decimal numeroAcordo, decimal multa, decimal juros, DateTime dataVencimento, decimal valorParcela, decimal? codErro, string cnpjEmpresaCobranca, int parcela, string sistema, string situacaoAluno, string tipoInadimplencia)
+        public async Task InserirBaixa(DateTime dataBaixa, decimal matricula, decimal numeroAcordo, decimal multa, decimal juros, DateTime dataVencimento, decimal valorParcela, int? codErro, string cnpjEmpresaCobranca, int parcela, string sistema, string situacaoAluno, string tipoInadimplencia)
         {
             await _repository.InserirBaixa(dataBaixa,
                                            matricula,
