+++ src/Tiradentes.CobrancaAtiva.Services/Services/ItensBaixasTipo3Service.cs
@@ -12,7 +12,7 @@ namespace Tiradentes.CobrancaAtiva.Services.Services
         {
             _repository = repository;
         }
-        public async Task InserirBaixa(DateTime dataBaixa, decimal matricula, decimal numeroAcordo, int parcela, DateTime dataPagamento, decimal valorPago, decimal? codErro)
+        public async Task InserirBaixa(DateTime dataBaixa, decimal matricula, decimal numeroAcordo, int parcela, DateTime dataPagamento, decimal valorPago, int? codErro)
         {
             await _repository.InserirBaixa(dataBaixa,
                                            matricula,
